Some code cleanup with elimination of numerous warnings

+++ BrainSimulator/PropertiesDlg.xaml.cs
@@ -3,19 +3,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 //  
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace BrainSimulator
 {
@@ -28,14 +16,14 @@ namespace BrainSimulator
         {
             InitializeComponent();
             if (MainWindow.theNeuronArray == null)
-            { Close();return; }
+            { Close(); return; }
 
             txtFileName.Text = MainWindow.currentFileName;
             txtFileName.ToolTip = MainWindow.currentFileName;
             txtRows.Text = MainWindow.theNeuronArray.rows.ToString("N0");
             txtColumns.Text = (MainWindow.theNeuronArray.arraySize / MainWindow.theNeuronArray.rows).ToString("N0");
             txtNeurons.Text = MainWindow.theNeuronArray.arraySize.ToString("N0");
-            MainWindow.theNeuronArray.GetCounts(out int synapseCount, out int neuronInUseCount);
+            MainWindow.theNeuronArray.GetCounts(out long synapseCount, out int neuronInUseCount);
             txtNeuronsInUse.Text = neuronInUseCount.ToString("N0");
             txtSynapses.Text = synapseCount.ToString("N0");
         }
